[geometry] Move client example binary and delete the dev folder
