--- geometry/render/dev/render_gltf_client/BUILD.bazel
@@ -1,65 +1,0 @@
-# -*- python -*-
-
-load(
-    "@drake//tools/skylark:drake_cc.bzl",
-    "drake_cc_binary",
-    "drake_cc_googletest",
-    "drake_cc_library",
-    "drake_cc_test",
-)
-load("@drake//tools/skylark:drake_py.bzl", "drake_py_unittest")
-load("//tools/skylark:py.bzl", "py_binary")
-load("//tools/lint:lint.bzl", "add_lint_tests")
-load("//tools/skylark:test_tags.bzl", "vtk_test_tags")
-
-package(default_visibility = ["//visibility:private"])
-
-# === test/ ===
-
-# - Server integration test targets. ------------------------------------------
-
-# NOTE: see geometry/render/dev/render_server/README.md for installation
-# instructions to run this server, flask is only installed for testing.
-# TODO(#16822): this test confirms `flask` can be imported, not run.  Remove in
-# the round trip tests.
-drake_py_unittest(name = "flask_import_test")
-
-filegroup(
-    name = "test_resources",
-    srcs = [
-        "test/4_color_texture.png",
-    ],
-)
-
-drake_cc_binary(
-    name = "run_simulation_and_render",
-    srcs = ["test/run_simulation_and_render.cc"],
-    add_test_rule = 1,
-    data = [
-        ":test_resources",
-        "//manipulation/models/ycb:models",
-    ],
-    test_rule_args = [
-        "--simulation_time=0.1",
-    ],
-    test_rule_tags = vtk_test_tags(),
-    deps = [
-        "//common:filesystem",
-        "//geometry:drake_visualizer",
-        "//geometry:scene_graph",
-        "//geometry/render_gltf_client:factory",
-        "//geometry/render_vtk",
-        "//multibody/parsing",
-        "//multibody/plant",
-        "//systems/analysis:simulator",
-        "//systems/framework:diagram",
-        "//systems/lcm:lcm_pubsub_system",
-        "//systems/primitives:constant_vector_source",
-        "//systems/sensors:image_to_lcm_image_array_t",
-        "//systems/sensors:image_writer",
-        "//systems/sensors:rgbd_sensor",
-        "@gflags",
-    ],
-)
-
-add_lint_tests()
