+++ geometry/render_gltf_client/test/4_color_texture.png
