+++ geometry/render_gltf_client/test/acceptance_test.py
@@ -35,31 +35,21 @@ MINIMAL_GLTF = """\
 """
 
 
-class TestGltfRenderServer(unittest.TestCase):
+class TestGltfRenderBinary(unittest.TestCase):
     def setUp(self):
         self.runfiles = CreateRunfiles()
 
-    def test_gltf_render_server_py(self):
-        """A minimal smoke test to ensure running the binary won't crash, e.g.,
-        the imports are correct.  It doesn't currently test receiving a flask
-        request and exercising `render_callback()`."""
+    def test_server_vtk_backend(self):
+        """Renders each type of images by invoking the vtk backend with test
+        arguments."""
 
-        server_py_binary = self.runfiles.Rlocation(
-            "drake/geometry/render_gltf_client/gltf_render_server"
-        )
-        subprocess.run([server_py_binary, "--test"])
-
-    def test_gltf_render_server_backend(self):
-        """Renders each type of images by invoking the backend binary with
-        test arguments."""
-
-        server_backend_binary = self.runfiles.Rlocation(
-            "drake/geometry/render_gltf_client/gltf_render_server_backend"
+        server_vtk_backend = self.runfiles.Rlocation(
+            "drake/geometry/render_gltf_client/server_vtk_backend"
         )
 
         # Note: `tmp_dir` will be cleaned up by bazel after running the test.
         tmp_dir = os.path.join(
-            os.environ.get("TEST_TMPDIR", "/tmp"), "gltf_render_server_test"
+            os.environ.get("TEST_TMPDIR", "/tmp"), "server_vtk_backend"
         )
         os.makedirs(tmp_dir, exist_ok=True)
         gltf_input_path = os.path.join(tmp_dir, "test.gltf")
@@ -72,7 +62,7 @@ class TestGltfRenderServer(unittest.TestCase):
         ):
             image_output_path = os.path.join(tmp_dir, f"{image_type}.{ext}")
             proc_args = [
-                server_backend_binary,
+                server_vtk_backend,
                 "--input_path",
                 gltf_input_path,
                 "--output_path",
@@ -103,3 +93,45 @@ class TestGltfRenderServer(unittest.TestCase):
 
             proc = subprocess.run(proc_args, capture_output=True)
             self.assertTrue(proc.returncode == 0)
+
+    def test_server_demo(self):
+        """A minimal smoke test to ensure running the server demo won't crash,
+        e.g., the imports are correct.  It doesn't currently test receiving a
+        flask request and exercising `render_callback()`."""
+
+        server_demo = self.runfiles.Rlocation(
+            "drake/geometry/render_gltf_client/server_demo"
+        )
+        subprocess.run([server_demo, "--test"])
+
+    def test_client_demo_client(self):
+        """A minimal smoke test to run the client demo that launches a Drake
+        simulation and a RenderEngineGltfClient for a short period of time."""
+        client_demo = self.runfiles.Rlocation(
+            "drake/geometry/render_gltf_client/client_demo"
+        )
+
+        # `client_demo` program defaults to instantiate RenderEngineGltfClient
+        # for rendering.
+        proc_args = [
+          client_demo,
+          "--simulation_time",
+          "0.1",
+        ]
+        subprocess.run(proc_args)
+
+    def test_client_demo_vtk(self):
+        """A minimal smoke test to run the client demo that launches a Drake
+        simulation and a RenderEngineVtk for a short period of time."""
+        client_demo = self.runfiles.Rlocation(
+            "drake/geometry/render_gltf_client/client_demo"
+        )
+
+        proc_args = [
+          client_demo,
+          "--render_engine",
+          "vtk",
+          "--simulation_time",
+          "0.1",
+        ]
+        subprocess.run(proc_args)
